+++ src/scripts/containers/filters/company-size.js
@@ -1,18 +1,8 @@
 import { setFilter, getState, subscribe } from "../../store/filter-store.js";
+import { COMPANY_SIZES } from "../../../constants/company-sizes.js";
 
 let selected = new Set();
-const allOptions = [
-  "0-1",
-  "2-10",
-  "11-50",
-  "51-200",
-  "201-500",
-  "501-1000",
-  "1001-5000",
-  "5001-10000",
-  "10001+",
-  "unknown",
-];
+const allOptions = COMPANY_SIZES;
 
 export function initCompanySizeFilter() {
   const container = document.getElementById("company-size-filter");